raftstore: extract more fsm::apply
